#SpringCloud Service for adding stock service and seata AT mode

+++ cloud-api-commons/src/main/java/com/springcloud/pojo/Payment.java
@@ -20,7 +20,7 @@ public class Payment implements Serializable {
     private String paymentSerial;
     private BigDecimal paymentPrice;
     private BigDecimal paymentTotalPrice;
-    private Long paymentNum;
+    private int paymentNum;
     private Timestamp updateTime;
     private Long productId;
